Final Covitionary Resource
This is the Covitionary Resource of frequent terms used in COVID-19 domain. The Covitionary resource has currently 12184 most frequent terms used in the COVID-19 domain literature. This resource has many potential applications such as creating ontological resources, automatic spell-checking in Covid domain, and classification of terms in pre-defined categories to name a few.